+++ php/Terminus/Endpoint.php
@@ -8,10 +8,10 @@ namespace Terminus;
  **/
 class Endpoint {
   public $patterns = array(
-    'deprecated' => 'https://%s/terminus.php?%s=%s',
-    'private'    => 'https://%s/api/%s/%s',
-    'public'     => 'https://%s/api/%s',
-    'login'      => 'https://%s/api/authorize',
+    'deprecated' => '%s://%s:%s/terminus.php?%s=%s',
+    'private'    => '%s://%s:%s/api/%s/%s',
+    'public'     => '%s://%s:%s/api/%s',
+    'login'      => '%s://%s:%s/api/authorize',
   );
 
   private $public_realms = array(
@@ -62,12 +62,20 @@ class Endpoint {
       $this->target = 'login';
     }
 
+    if (!isset($args['protocol']) || ($args['protocol'] == '')) {
+      $args['protocol'] = TERMINUS_PROTOCOL;
+    }
+
     if (!isset($args['host']) || ($args['host'] == '')) {
       $args['host'] = TERMINUS_HOST;
     }
 
+    if (!isset($args['port']) || ($args['port'] == '')) {
+      $args['port'] = TERMINUS_PORT;
+    }
+
     //A substiution array to pass to the vsprintf
-    $substitutions = array($args['host'], $args['realm']);
+    $substitutions = array($args['protocol'], $args['host'], $args['port'], $args['realm']);
     if (isset($args['uuid']) && $args['uuid'] != 'public') {
       array_push($substitutions, $args['uuid']);
     }