+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 ### Fixed
 - `sites list` no longer capitalizes membership UUIDs. (#642)
 
+### Added
+- Environment variables `TERMINUS_PORT` and `TERMINUS_PROTOCOL` now enabled. (#643)
+
 ##[0.9.1] - 2015-10-27
 ### Fixed
 - `site backups get` will now find and retrieve backups properly. (#632)