+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 #Change Log
 All notable changes to this project starting with the 0.6.0 release will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
-##Master
+##[0.9.2] - 2015-10-29
 ### Fixed
 - `sites list` no longer capitalizes membership UUIDs. (#642)